Added scheduled external api call

+++ src/main/resources/db/changelog/001_initial_schema.sql
@@ -2,9 +2,9 @@ CREATE TABLE public.weather_info
 (
     id bigserial NOT NULL,
     humidity smallint,
-    pressure_millibars smallint,
-    temp_celsius smallint,
-    temp_fahrenheit smallint,
+    pressure_millibars double precision,
+    temp_celsius double precision,
+    temp_fahrenheit double precision,
     wind_speed_ms double precision,
     requested_at timestamp,
     city character varying(255) COLLATE pg_catalog."default",
